Buffer process output on a separate thread.
On MacOS the pipe for a PTY's STDOUT is 1024 bytes. This has the effect
of slowing down all processes in eshell because the 1024 byte buffer
is consumed on the main event loop and includes parsing and
calling the process-filter for every single 1024 byte buffer.

Consuming stdout on a background thread allows the 1024 byte buffers
to be coallesced into a single buffer that the main event loop can
then use. This has the effect of increasing the size of the buffer
that the process-filter can be called with, which reduces the amount
of invalid intermediate states that are parsed when using
autocompletion frameworks.