Switch to using a combine write cache and increase the cache size
Also, set the layers to opaque, since additional optimizations can be
made with the CALayer being opaque. The combine write cache is useful
because the number of writes far exceeds the reads. Anecdotally, it
seems to work better than before.

Increasing the cache size helps in some worst case scenarios where
there would be contention over the the surface. Increasing the cache
size avoids that at the expense of some wasted memory.